Sept-06 Update
revisions per email:

Hi Arno, I know that David wants to get the site out into the public as soon as possible. So let's take next steps to get there quickly. Here are some final layout details to address:

Opening Section / Hero Banner

1) avoid .png/.jpg for any icons or other small pics that need to be razor sharp; with very few exception all browsers can handle .svgs. So, replace /img/penta_icon.png with /img/logo.svg (loads faster too) I do have the Penta logo by itself as a .svg (all are attached and can be easily resized/modified in illustrator)

2) the opening text font size seems small, consider setting the <h4> to <h3>; I would also change the css so that the block is not as narrow (since it is a grand opening title) from "col-md-6 col-md-offset-3" to "col-md-8 col-md-offset-2". That extra space makes it much easier to read. Last css polish is to add a "margin-top: 0; and text-align:justify;" to the <div> so the marquee text is closer to the "PENTA GLOBAL" title and has crisp left/right edges making it much easier to read

3) for the opening text please change from a passive voice to active; suggest something like:

"Delivering the potential of blockchain technology with infrastructure for individuals and corporations to interconnect and interoperate" or "Delivering the potential of blockchain technology with infrastructure for the world to interconnect and interoperate". I am still not sure if this should be in titlecase or written as a simple sentence, probably titlecase is best and consistent.
4) the black 'connect with us' band seems overwhelming, try using rgba(50, 51, 66, 1.0) where the last number (1.0) controls opacity. Another idea, that <div> can be made part of the hero banner, letting the animation show through from behind a bit, which would neatly blend everything together.

5) I am attaching a screenshot of the hero section restyled.

We Believe in Great Ideas Section

1) the blockchain rice background is a nice idea, better than the laptop or the iphone/astronaut, but in using it we now need to restyle the text a bit so it does not get overwhelmed. Try changing the container 100% width down to 66% for the wrapper that holds the title and text, from: class="col-md-12 col-sm-12 " to class="col-md-8 col-sm-12". Then add to the font size in the airspace.css#438 ("#feature p" selector) with: "font-size: 18px;"

2) the penta_rise_bg.jpg is a whopping 1M in size. Attached is an optimized version that comes in at 158kb. Still big for my taste, but will do.

Video

1) On my desktop machine the embedded video title/key frame is blurred; let me suggest the video be edited using something like iMovie so that a title can be easily added. Not that I do not enjoy looking at a blurred picture of David :) but something a bit sharper will look better.

2) the css for the section#youtube should have its background color set to #3e2966

3) the "row" <div> that contains the video should have some padding added to its styling so the video does not bump up against the top/bottom of the section: "padding: 70px 0;"

4) I have attached a screenshot of the section with the added padding so you can see how the layout styling works and why I suggest editing the video some more.

Roadmap

1) I would move the Roadmap section out of the "About Penta" section and into the "Technology Without Borders" one; that will balance things out a bit

2) Attached is a screenshot illustrating the final layout with the Roadmap moved and the text restyled with its color set to #efefef

3) The text itself, both the description and the timeline labels can be shortened. For example: "Penta Network Testnet release" to "TestNet Release" or "Official Penta Launch" to "Penta Launch" etc.

4) For the descriptive text suggest something like:

"Penta Network launched January 2018 in Singapore. The goal: a new-generation of blockchain network. Since then Penta has been working hard with research to build Penta DLOS, Penta Blockchain, the high performance DApp platform and Penta connector. We are delivering blockchain as a universally-benefiting technology, and ultimately a platform to interconnect all the world’s blockchain ecosystems."

Ambassador program

1) I would make sure all the images are the same size; this holds true for the projects section too

Menubar

1) I will increase the size of the world icon to 24px; and add a "margin-top: -4px" to the wrapping <a> so the icon is vertically aligned

There are a few other bits that could use some polish, but I know the above is a lot to handle so let's start with them. If you are in agreement I can certainly handle all these updates for you in the next 24 hours if you push them to the/a development branch in GitHub for review and merging with the 'master'.

With regards deployment, there are a number of options, depending on what you would like to accomplish:

1) are you looking to automate posting changes to github and then to AWS?

2) are you looking to automate posting from github to AWS? If this is the case, Amazon has some tools to help:

https://docs.aws.amazon.com/codedeploy/latest/userguide/tutorials-github.html

The basic workflow I had in mind is that we give edit privileges to all those folks who will be posting to the site (David, Roy, ...) They install github desktop (with our help) and pull the site onto their systems. Their edits would be to the development branch, or possibly to their own branch. They then push their changes back to Github. Whoever is tasked with maintenance then pulls any of these changes (to the dev branch) for testing on their local machine via a 'Hugo' command. So they need Hugo on their computer, BUT 'posters' do not. Posters just need to know how to use the basics of markdown, a text editor and Github Desktop. Once testing is complete the maintainer pushes the new version of the dev branch back to Github. That person, or whoever is the lead gets a pull request to 'merge' the test branch into the "master" (production) one. When that is done (should be less than a minute) they pull the new 'master' branch to their local machine and rebuild using Hugo. The new site is then pushed back up to Github and deployed to our Penta webserver using any tool we wish. Alternatively, you could also deploy directly from your local machine to the webserver and then update GitHub.

The main differences between this and the digitalocean tutorial is that we are simplifying by keeping any Hugo building local, rather than on a remote server. We also are utilizing GitHub desktop as a workflow manager, which in the future is probably the best way to go as we can integrate with tools like Trello. If we do not want to burden people with GitHub then they can just edit a .md file (we can provide templates for that) and send on via WeChat or email. The maintainers can then add any new files to the repo dev branch and proceed that way.

I know this is centralizing the process and for a blockchain company goes against an ethos of decentralization, but maybe a project for the future is a blockchain based code versioning system :)

Hope this helps, Steve